Refactored build:ci:configure command
