+++ src/ServiceProviders/RepositoryProviders/GitProvider.php
@@ -58,4 +58,13 @@ interface GitProvider extends ProviderInterface
      * @return string URL to target project.
      */
     public function projectURL($target_project);
+
+    /**
+     * Git Commit URL to add as a Git remote.
+     *
+     * @param string $target_project Project to generate Git URL for.
+     *
+     * @return string URL to target project.
+     */
+     public function gitCommitURL($target_project);
 }
